+++ modules/elementary_functions/builtin/cpp/Gateway.cpp
@@ -69,6 +69,7 @@
 #include "prodBuiltin.hpp"
 #include "modBuiltin.hpp"
 #include "absBuiltin.hpp"
+#include "repmatBuiltin.hpp"
 //=============================================================================
 using namespace Nelson;
 //=============================================================================
@@ -130,6 +131,7 @@ static const nlsGateway gateway[] =
     { "prod", Nelson::ElementaryFunctionsGateway::prodBuiltin, 1, 3 },
     { "mod", Nelson::ElementaryFunctionsGateway::modBuiltin, 1, 2 },
 	{ "abs", Nelson::ElementaryFunctionsGateway::absBuiltin, 1, 1 },
+    { "repmat", Nelson::ElementaryFunctionsGateway::repmatBuiltin, 1, -1 },
 };
 //=============================================================================
 NLSGATEWAYFUNC(gateway)